feat: add platform-agnostic mining verification and lifecycle improvements (#7)
**Key Changes:**

- Added a cross-platform GitHub Action to verify actual mining
functionality and catch CPU mining regressions (e.g., hash rate stuck at
0 H/s)
- Refactored mining client and server for improved lifecycle management,
platform-agnostic self-deletion, and robust shutdown signaling
- Overhauled workflow pipelines to include mining verification jobs on
Windows, Linux, and macOS, ensuring true block mining on all platforms
- Hardened API and pool server for concurrent use, graceful shutdown,
and stricter type consistency for difficulty and throttle parameters

**Added:**

- `.github/actions/verify-mining`: New composite GitHub Action (with
`action.yml`, `verify.sh`, `verify.ps1`, and detailed `README.md`) for
end-to-end mining verification, testing real block production and hash
rate on all OSes
- Mining verification jobs in `test-and-build.yaml` and
`windows-miner-test.yaml`, running after code builds and uploading test
artifacts for each platform
- Markdown template (`tools/templates/report_template.tmpl`) for
structured, detailed impact assessment report generation in the
reporting tool

**Changed:**

- Client:
- Improved mining client documentation, clarified mining modes,
configuration, and control features
- Implemented robust, platform-aware self-deletion logic using safe path
handling and deletion scripts for Windows
- Added atomic shutdown signaling, file-based shutdown triggers, and
improved graceful shutdown across all platforms
- Introduced `clampToInt32` for safe conversion of hardware attributes
sent via protobuf; fully tested edge cases
- Enhanced miner and mining loop structures for hybrid, GPU, and CPU
mining, with clearer lifecycle and shutdown management
- Server:
- Refactored API server and WebSocket hub for context-based, cancellable
goroutine management, supporting graceful shutdown
- All server API, logger, and pool types now use consistent `int32` for
difficulty and throttle values, aligning with protobuf and client
- Pool and API server now accept and propagate context for cancellation,
and expose explicit `Shutdown()` methods
- Hardened logging to atomic writes, robust error handling, and improved
concurrency safety
  - Improved documentation and comments for maintainability and clarity
- Workflows:
- Updated all CI jobs to pin GitHub Action versions with full commit
hashes for security
  - All artifact download and upload steps now use fixed versions
  - Test jobs now upload all mining verification logs as build artifacts
- Summary jobs report the result of mining verification for all
platforms
- Reporting tool:
- Generates markdown using the new template, with improved filename
validation and safe file handling

**Removed:**

- Manual, inline mining regression checks from individual workflow jobs,
now superseded by the platform-agnostic `verify-mining` action
- Redundant or outdated comments in workflow YAML and client/server
code, replaced by structured doc comments and improved in-code
documentation
- Unused or duplicate code for error reporting in API handlers, now
unified under a single `writeJSON` utility function

+++ server/grpc_server.go
@@ -91,7 +91,7 @@ func (s *MiningPoolServer) GetWork(ctx context.Context, req *pb.WorkRequest) (*p
 		BlockIndex:   block.Index,
 		PreviousHash: block.PreviousHash,
 		Data:         block.Data,
-		Difficulty:   int32(s.pool.blockchain.Difficulty),
+		Difficulty:   s.pool.blockchain.Difficulty,
 		Timestamp:    block.Timestamp,
 	}, nil
 }
@@ -192,7 +192,7 @@ func (s *MiningPoolServer) Heartbeat(ctx context.Context, req *pb.MinerStatus) (
 		Active:             true,
 		Message:            "Heartbeat received",
 		ShouldMine:         shouldMine,
-		CpuThrottlePercent: int32(cpuThrottle),
+		CpuThrottlePercent: cpuThrottle,
 	}, nil
 }
